refactor(format): extract json export owner
